System: move error and exception handling to an ErrorHandler class
