+++ resources/templates/error.twig.html
@@ -1,0 +1,19 @@
+{#<!--
+Gibbon, Flexible & Open School System
+Copyright (C) 2010, Ross Parker
+
+This is a Gibbon template file, written in HTML and Twig syntax.
+For info about editing, see: https://twig.symfony.com/doc/2.x/
+-->#}
+{% extends "index.twig.html" %}
+
+{% block page %}
+    <h1>
+        {{ __('Oh no!') }}
+    </h1>
+    <p>
+        {{ __('Something has gone wrong: the Gibbons have escaped!') }}<br/>
+        <br/>
+        {{ __('An error has occurred. This could mean a number of different things, but generally indicates that you have a misspelt address, or are trying to access a page that you are not permitted to access.') }} {{ __('If you cannot solve this problem by retyping the address, or through other means, please contact your system administrator.') }}<br/>
+    </p>
+{% endblock %}
